move modeling.py and modeling_nv.py to transformers
