+++ slm/pipelines/examples/contrastive_training/evaluation/eval_mteb.py
@@ -14,16 +14,12 @@
 
 import argparse
 import logging
-import sys
 
 import mteb
-import paddle
 from mteb import MTEB
 
 from paddlenlp.peft import LoRAConfig, LoRAModel
-from paddlenlp.transformers import AutoModel, AutoModelForCausalLM, AutoTokenizer
-from paddlenlp.transformers.llm_embed.modeling import BiEncoderModel
-from paddlenlp.transformers.nv_embed.modeling_nv import NVEncodeModel
+from paddlenlp.transformers import AutoTokenizer, BiEncoderModel, NVEncodeModel
 
 
 class MTEB_EvalModel:
